Fixed the four example philosophers.  Closes #16 and #17.
Refactored a bunch of the article parsing code.  It reads better,
works better, and is much more robust in terms of guessing whether a
person is alive or not.  There was a logic problem in the alive?
method and it did not consider the persondata template contained in
all articles of people.  The project is finally ready for
disambiguations.